Updated all loot filters for patch 3.7.0 / Legion
As usual, the game won't accept these new filter parameters until the 
new patch is out tonight, and if there are typos in here or if the news 
post had issues, they will be corrected the first time I log into the 
game after the patch comes out. Enjoy Legion!